refactor: definition.ts path
