+++ src/features/groups/components/ContainerList.tsx
@@ -1,7 +1,6 @@
 import { Card, Icon } from '@/components/ui';
 import { fetchContainerList } from '@/lib/api/data';
-
-import { IContainer } from '../types/definition';
+import { IContainer } from '@/types/definition';
 
 export const ContainerList = async ({ id }: { id: string }) => {
   const containers: IContainer[] = await fetchContainerList(id);
